backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-genai>=1.0.0
Pillow==10.1.0
python-dotenv==1.0.0
requests>=2.32.2
pydantic>=2.9